+++ tests/test_dbbase/base.py
@@ -1,8 +1,6 @@
 # test/test_dbbase/base.py
 from . import DBBaseTestCase
 
-import sqlalchemy
-
 
 class TestDBBaseClass(DBBaseTestCase):
     """
@@ -17,7 +15,6 @@ class TestDBBaseClass(DBBaseTestCase):
 
         """
         DB = self.dbbase.DB
-        Model = self.dbbase.model.Model
 
         # test defaults
         # must have config
@@ -40,6 +37,176 @@ class TestDBBaseClass(DBBaseTestCase):
         # echo
 
         # test self.attributes
+
+    def test_doc_table(self):
+        """test_doc_table
+
+        This test ensures that only columns, not relations are passed to
+        process_expression.
+        """
+        db = self.db
+
+        class Table1(db.Model):
+            __tablename__ = 'table1'
+            id = db.Column(db.Integer, primary_key=True)
+            table2 = db.relationship(
+                "Table2", backref="table1", lazy='immediate')
+
+        class Table2(db.Model):
+            __tablename__ = 'table2'
+            id = db.Column(db.Integer, primary_key=True)
+
+            table1_id = db.Column(db.Integer, db.ForeignKey('table1.id'))
+
+        class Table0(db.Model):
+            __tablename__ = 'table0'
+            id = db.Column(db.Integer, primary_key=True)
+
+            table1_id = db.Column(db.Integer, db.ForeignKey('table1.id'))
+
+        db.create_all()
+
+        # Table1
+        self.assertDictEqual(
+            {
+                'Table1': {
+                    'type': 'object',
+                    'properties': {
+                        'id': {
+                            'name': 'id',
+                            'type': 'integer',
+                            'format': 'int32',
+                            'primary_key': True,
+                            'nullable': False,
+                            'info': {}
+                        }
+                    }
+                }
+            },
+            db.doc_table(Table1)
+        )
+
+        # Table2
+        self.assertDictEqual(
+            {
+                'Table2': {
+                    'type': 'object',
+                    'properties': {
+                        'id': {
+                            'name': 'id',
+                            'type': 'integer',
+                            'format': 'int32',
+                            'primary_key': True,
+                            'nullable': False,
+                            'info': {}},
+                        'table1_id': {
+                            'name': 'table1_id',
+                            'type': 'integer',
+                            'format': 'int32',
+                            'nullable': True,
+                            'foreign_key':
+                                'table1.id',
+                                'info': {}
+                        }
+                    }
+                }
+            },
+            db.doc_table(Table2)
+        )
+
+        # Test doc_tables
+        doc = db.doc_tables()
+        self.assertDictEqual(
+            {
+                "definitions": {
+                    "Table0": {
+                        "type": "object",
+                        "properties": {
+                            "table1_id": {
+                                "name": "table1_id",
+                                "type": "integer",
+                                "format": "int32",
+                                "nullable": True,
+                                "foreign_key": "table1.id",
+                                "info": {}
+                            },
+                            "id": {
+                                "name": "id",
+                                "type": "integer",
+                                "format": "int32",
+                                "primary_key": True,
+                                "nullable": False,
+                                "info": {}
+                            }
+                        }
+                    },
+                    "Table1": {
+                        "type": "object",
+                        "properties": {
+                            "id": {
+                                "name": "id",
+                                "type": "integer",
+                                "format": "int32",
+                                "primary_key": True,
+                                "nullable": False,
+                                "info": {}
+                            }
+                        }
+                    },
+                    "Table2": {
+                        "type": "object",
+                        "properties": {
+                            "table1_id": {
+                                "name": "table1_id",
+                                "type": "integer",
+                                "format": "int32",
+                                "nullable": True,
+                                "foreign_key": "table1.id",
+                                "info": {}
+                            },
+                            "id": {
+                                "name": "id",
+                                "type": "integer",
+                                "format": "int32",
+                                "primary_key": True,
+                                "nullable": False,
+                                "info": {}
+                            }
+                        }
+                    }
+                }
+            },
+            doc
+        )
+
+        # correct number
+        self.assertEqual(3, len(doc['definitions'].keys()))
+
+        # does it sort
+        self.assertListEqual(
+            ['Table0', 'Table1', 'Table2'],
+            [key for key in doc['definitions'].keys()]
+        )
+
+        # does class list work
+        doc = db.doc_tables(class_list=[Table2])
+        self.assertEqual(1, len(list(doc['definitions'].keys())))
+        self.assertEqual('Table2', list(doc['definitions'].keys())[0])
+
+        # does it pass through camel flag
+        doc = db.doc_tables(class_list=[Table2], to_camel_case=True)
+        self.assertIn(
+            'table1Id',
+            list(doc['definitions']['Table2']['properties'].keys())
+        )
+
+        # does it pass through column props filter
+        doc = db.doc_tables(column_props=['name', 'primary_key'])
+
+        self.assertNotIn(
+            'type',
+            list(doc['definitions']['Table2']['properties'].keys())
+        )
 
     def test_create_engine(self):
         pass